fix stats counter, move to separate file
